# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.errorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  model.AcademicYear:
    properties:
      created_at:
        type: string
      ends_on:
        type: string
      id:
        type: integer
      school_id:
        type: integer
      starts_on:
        type: string
      updated_at:
        type: string
      year_name:
        type: string
    type: object
  model.AuditLog:
    properties:
      action:
        type: string
      created_at:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      metadata:
        type: object
      record_id:
        type: integer
      resource_type:
        type: string
      school_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  model.Permission:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  model.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/model.Permission'
        type: array
      status:
        $ref: '#/definitions/model.RoleStatus'
      updated_at:
        type: string
    type: object
...
  service.CreateAcademicYearInput:
    properties:
      ends_on:
        example: "2027-06-30"
        type: string
      starts_on:
        example: "2026-09-01"
        type: string
      year_name:
        example: 2026-2027
        maxLength: 20
        type: string
    required:
    - ends_on
    - starts_on
    - year_name
    type: object
...
  service.ReplaceRolePermissionsInput:
    properties:
      permission_ids:
        items:
          type: integer
        type: array
    required:
    - permission_ids
    type: object
...
  service.UpdateAcademicYearInput:
    properties:
      ends_on:
        example: "2027-06-30"
        type: string
      starts_on:
        example: "2026-09-01"
        type: string
      year_name:
        example: 2026-2027
        maxLength: 20
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for Kobciye School.
  title: Kobciye School API
  version: "1.0"
paths:
  /academic-years:
    get:
      parameters:
      - description: Required for SuperAdmin
        in: header
        name: X-School-ID
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.AcademicYear'
            type: array
      security:
      - BearerAuth: []
      summary: List academic years for the resolved school
      tags:
      - foundation
    post:
      consumes:
      - application/json
      parameters:
      - description: Required for SuperAdmin
        in: header
        name: X-School-ID
        type: integer
      - description: Academic year details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.CreateAcademicYearInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.AcademicYear'
      security:
      - BearerAuth: []
      summary: Create an academic year for the resolved school
      tags:
      - foundation
  /academic-years/{id}:
    delete:
      description: Permanently deletes an academic year only when no dependent records
        use it.
      parameters:
      - description: Required for SuperAdmin
        in: header
        name: X-School-ID
        type: integer
      - description: Academic year ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete an academic year
      tags:
      - foundation
    patch:
      consumes:
      - application/json
      parameters:
      - description: Required for SuperAdmin
        in: header
        name: X-School-ID
        type: integer
      - description: Academic year ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.UpdateAcademicYearInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AcademicYear'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Update an academic year
      tags:
      - foundation
  /audit-logs:
    get:
      parameters:
      - description: Optional SuperAdmin filter
        in: query
        name: school_id
        type: integer
      - description: Page size, maximum 100
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.AuditLog'
            type: array
      security:
      - BearerAuth: []
      summary: List append-only audit logs
      tags:
      - foundation
...
  /permissions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Permission'
            type: array
      security:
      - BearerAuth: []
      summary: List permissions
      tags:
      - foundation
  /roles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Role'
            type: array
      security:
      - BearerAuth: []
      summary: List roles and their permissions
      tags:
      - foundation
    post:
      consumes:
      - application/json
      parameters:
      - description: Role and permissions
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.CreateRoleInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Role'
      security:
      - BearerAuth: []
      summary: Create a custom role
      tags:
      - foundation
  /roles/{id}:
    delete:
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Deactivate a custom role
      tags:
      - foundation
    patch:
      consumes:
      - application/json
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.UpdateRoleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Role'
      security:
      - BearerAuth: []
      summary: Update a custom role
      tags:
      - foundation
  /roles/{id}/permissions:
    put:
      consumes:
      - application/json
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Complete permission assignment
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.ReplaceRolePermissionsInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Role'
      security:
      - BearerAuth: []
      summary: Replace a custom role's permissions
      tags:
      - foundation
  /schools:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.School'
            type: array
      security:
      - BearerAuth: []
      summary: List schools
      tags:
      - foundation
    post:
      consumes:
      - application/json
      parameters:
      - description: School details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.CreateSchoolInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.School'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a school
      tags:
      - foundation
  /schools/{id}:
    delete:
      description: 'Safe delete: sets the school to inactive instead of removing historical
        data.'
      parameters:
      - description: School ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Deactivate a school
      tags:
      - foundation
    get:
      parameters:
      - description: School ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.School'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a school by ID
      tags:
      - foundation
    patch:
      consumes:
      - application/json
      description: Partially updates a school. Setting status to inactive immediately
        blocks its school accounts.
      parameters:
      - description: School ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.UpdateSchoolInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.School'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a school
      tags:
      - foundation
  /users:
    get:
      parameters:
      - description: Optional SuperAdmin filter
        in: query
        name: school_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
      security:
      - BearerAuth: []
      summary: List users in the permitted school scope
      tags:
      - foundation
    post:
      consumes:
      - application/json
      parameters:
      - description: User details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.CreateUserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
      security:
      - BearerAuth: []
      summary: Create a school user
      tags:
      - foundation
...
  /users/{id}/password-reset-token:
    post:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/service.PasswordResetTokenResult'
      security:
      - BearerAuth: []
      summary: Create a one-time password reset token for a managed user
      tags:
      - foundation
  /users/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New account status
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/service.UpdateUserStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      security:
      - BearerAuth: []
      summary: Disable or unlock a user account
      tags:
      - foundation
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"